+++ apps/server/src/app.ts
@@ -2,12 +2,8 @@ import cors from 'cors'
 import express, { Request, Response, Express } from 'express'
 import { createExpressMiddleware as createMid } from '@trpc/server/adapters/express'
 import { appRouter } from './trpc'
-import {
-  createOpenApiExpressMiddleware,
-  generateOpenApiDocument,
-} from 'trpc-openapi'
-import * as yaml from 'yaml'
-import * as fs from 'fs'
+import { createOpenApiExpressMiddleware } from 'trpc-openapi'
+import { trpcDocumentation } from './openapi'
 
 const corsOpts = {
   origin: [
@@ -34,29 +30,10 @@ export function getApp(): Express {
   /**
    * run auto generate
    */
-  trpcDocumentation()
+  trpcDocumentation(appRouter)
   /** register root route */
   app.get('/', (_req: Request, res: Response) => {
     res.status(200).send('modtree server is running')
   })
   return app
-}
-
-function trpcDocumentation() {
-  const openApiDocument = generateOpenApiDocument(appRouter, {
-    title: 'modtree API',
-    version: '1.0.0',
-    baseUrl: 'http://localhost:8080/api',
-  })
-
-  // YAML equivalent
-  const fileContent = yaml.stringify(openApiDocument)
-
-  const location = `${__dirname}/../../../apps/docs/public/openapi-docs.yml`
-
-  fs.writeFileSync(location, fileContent, {
-    encoding: 'utf-8',
-  })
-
-  console.debug(`Written OpenAPI docs to ${location}.`)
 }
